Backend/requirements.txt
fastapi
uvicorn[standard]
aiosqlite
python-jose[cryptography]
bcrypt
python-multipart
slowapi
python-dotenv
email-validator
aiosmtplib
